HHH-6819 Moving CommandExtractorService, MultiLineImportFileTest and SingleLineImportFileTest out of the matrix tests. The tests don't need to run against multiple dbs and the multi line tests are not portable as they are atm.
Also splitting properly source files from resources
