--- hibernate-core/src/matrix/java/org/hibernate/test/importfile/Dog.hbm.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-	"http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-
-<!-- 
-
-  Test for import files.
-     
--->
-
-<hibernate-mapping package="org.hibernate.test.importfile">
-    <class name="Dog" table="dog">
-        <id name="id"/>
-        <many-to-one name="master" column="master_fk"/>
-    </class>
-</hibernate-mapping>
