--- hibernate-core/src/matrix/java/org/hibernate/test/importfile/Human.hbm.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC 
-	"-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-	"http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-
-<!-- 
-
-  Test for import files.
-     
--->
-
-<hibernate-mapping package="org.hibernate.test.importfile">
-    <class name="Human" table="human">
-        <id name="id"/>
-        <property name="firstname" column="fname"/>
-        <property name="lastname" column="lname"/>
-    </class>
-</hibernate-mapping>
