+++ hibernate-core/src/test/resources/org/hibernate/test/fileimport/Dog.hbm.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0"?>
+<!DOCTYPE hibernate-mapping PUBLIC
+        "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
+        "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
+
+<hibernate-mapping package="org.hibernate.test.fileimport">
+    <class name="Dog" table="dog">
+        <id name="id"/>
+        <many-to-one name="master" column="master_fk"/>
+    </class>
+</hibernate-mapping>
