+++ hibernate-core/src/test/resources/org/hibernate/test/fileimport/Human.hbm.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<!DOCTYPE hibernate-mapping PUBLIC
+        "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
+        "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
+
+<hibernate-mapping package="org.hibernate.test.fileimport">
+    <class name="Human" table="human">
+        <id name="id"/>
+        <property name="firstname" column="fname"/>
+        <property name="lastname" column="lname"/>
+    </class>
+</hibernate-mapping>
