+++ hibernate-core/src/test/resources/org/hibernate/test/fileimport/dogs.sql
