+++ hibernate-core/src/test/resources/org/hibernate/test/fileimport/humans.sql
